# public/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/matchmaking/quickplay": {
      "post": {
        "description": "Join queue for Rock Paper Scissors match",
        "operationId": "MatchmakingController_quickplay",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Joined queue or match created"
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Join matchmaking queue",
        "tags": [
          "Matchmaking"
        ]
      }
    },
    "/matchmaking/match/{id}/audit": {
      "get": {
        "description": "Returns financial audit for match",
        "operationId": "MatchmakingController_audit",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get match audit log",
        "tags": [
          "Matchmaking"
        ]
      }
    },
    "/matchmaking/ticket/{id}": {
      "get": {
        "operationId": "MatchmakingController_ticket",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Matchmaking"
        ]
      }
    },
    "/matchmaking/ticket/{id}/fallback": {
      "post": {
        "operationId": "MatchmakingController_fallback",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Matchmaking"
        ]
      }
    },
    "/matchmaking/match/{id}": {
      "get": {
        "description": "Returns match state by ID",
        "operationId": "MatchmakingController_match",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get match details",
        "tags": [
          "Matchmaking"
        ]
      }
    },
    "/matchmaking/match/{id}/move": {
      "post": {
        "description": "Submit ROCK, PAPER or SCISSORS move",
        "operationId": "MatchmakingController_move",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitMoveDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Submit move",
        "tags": [
          "Matchmaking"
        ]
      }
    },
    "/matchmaking/cleanup-orphaned": {
      "post": {
        "operationId": "MatchmakingController_cleanupOrphaned",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Matchmaking"
        ]
      }
    },
    "/matchmaking/test-create-orphaned": {
      "post": {
        "operationId": "MatchmakingController_testCreateOrphaned",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Matchmaking"
        ]
      }
    },
    "/matchmaking/active": {
      "get": {
        "description": "Check if user is in queue or active match",
        "operationId": "MatchmakingController_getActiveState",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get active state",
        "tags": [
          "Matchmaking"
        ]
      }
    },
    "/matchmaking/online": {
      "get": {
        "operationId": "MatchmakingController_getOnlineCount",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Matchmaking"
        ]
      }
    },
    "/matchmaking/history": {
      "get": {
        "description": "Returns user's match history",
        "operationId": "MatchmakingController_getMatchHistory",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Get match history",
        "tags": [
          "Matchmaking"
        ]
      }
    },
    "/matchmaking/test/force-match": {
      "post": {
        "operationId": "MatchmakingController_forceMatch",
        "parameters": [
          {
            "name": "authorization",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Matchmaking"
        ]
      }
    },
...
    "/leaderboard": {
      "get": {
        "description": "Returns top players sorted by specified category",
        "operationId": "LeaderboardController_getLeaderboard",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Sort category",
            "schema": {
              "enum": [
                "wins",
                "winRate",
                "profit",
                "streak",
                "biggestWin"
              ],
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of results (default: 10, max: 100)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Offset for pagination (default: 0)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Leaderboard entries"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get leaderboard by category",
        "tags": [
          "Leaderboard"
        ]
      }
    },
    "/leaderboard/global": {
      "get": {
        "description": "Returns top players from all categories for main page",
        "operationId": "LeaderboardController_getGlobalLeaderboard",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of results per category (default: 5)",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Leaderboard by all categories"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get global leaderboard",
        "tags": [
          "Leaderboard"
        ]
      }
    },
    "/leaderboard/me": {
      "get": {
        "description": "Returns authenticated user's rank in leaderboard",
        "operationId": "LeaderboardController_getMyPosition",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Category",
            "schema": {
              "enum": [
                "wins",
                "winRate",
                "profit",
                "streak",
                "biggestWin"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User position"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user position",
        "tags": [
          "Leaderboard"
        ]
      }
    },
    "/leaderboard/user/{userId}": {
      "get": {
        "description": "Returns specific user's position",
        "operationId": "LeaderboardController_getUserPosition",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "Category",
            "schema": {
              "enum": [
                "wins",
                "winRate",
                "profit",
                "streak",
                "biggestWin"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User position"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get user position",
        "tags": [
          "Leaderboard"
        ]
      }
    },
    "/leaderboard/categories": {
      "get": {
        "operationId": "LeaderboardController_getCategories",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of categories"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get available leaderboard categories",
        "tags": [
          "Leaderboard"
        ]
      }
    }
  },
  "info": {
    "title": "WagerPlay API",
    "description": "Rock Paper Scissors betting game API",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "SubmitMoveDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}